Move error ID generator to error package
