+++ .changeset/add-error-id-generator.md
@@ -1,0 +1,5 @@
+---
+'@prairielearn/error': minor
+---
+
+Add a shared generator for error correlation IDs.
