+++ packages/trpc/package.json
@@ -29,8 +29,7 @@
     "react-bootstrap": "3.0.0-beta.5"
   },
   "dependencies": {
-    "@types/react": "^19.2.18",
-    "es-toolkit": "^1.50.0"
+    "@types/react": "^19.2.18"
   },
   "devDependencies": {
     "@prairielearn/tsconfig": "workspace:^",
